requirements.txt
annotated-types==0.6.0
anyio==4.2.0
apache-beam==2.53.0
asttokens==2.4.1
attrs==23.2.0
azure-core==1.29.6
azure-data-tables==12.4.4
azure-functions==1.18.0
azure-storage-blob==12.19.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.0
crcmod==1.7
cryptography==41.0.7
cx-Oracle==8.3.0
debugpy==1.8.0
decorator==5.1.1
dill==0.3.1.1
dnspython==2.4.2
docopt==0.6.2
exceptiongroup==1.2.0
executing==2.0.1
fastavro==1.9.2
fasteners==0.19
grpcio==1.60.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.2
ipykernel==6.28.0
ipython==8.12.3
isodate==0.6.1
jedi==0.19.1
Js2Py==0.74
jsonpickle==3.0.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.0
jupyter-core==5.6.0
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.24.4
objsize==0.6.1
orjson==3.9.10
packaging==23.2
pandas==2.0.3
paramiko==3.4.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pkgutil-resolve-name==1.3.10
platformdirs==4.1.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.5.3
pydantic-core==2.14.6
pydot==1.4.2
pygments==2.17.2
pyjsparser==2.7.1
pymongo==4.6.1
PyNaCl==1.5.0
pyodbc==5.1.0
pyparsing==3.1.1
pysftp==0.2.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.2
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rpds-py==0.16.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
typing-extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
wcwidth==0.2.12
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0
